% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.utils.R
\docType{package}
\name{fc.utils-package}
\alias{fc.utils}
\alias{fc.utils-package}
\title{fc.utils}
\description{
Utility functions used both in the construction of the fc-dataset and
analysis of frozen conflict onsets.
}
\author{
\strong{Maintainer}: Joshua Krusell \email{joshua.krusell@gu.se}

}
\keyword{internal}
