% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_msg}
\alias{log_msg}
\alias{info}
\alias{warn}
\title{Formatted logging}
\usage{
log_msg(type = c("INFO", "WARN", "ERROR"), ...)

info(...)

warn(...)
}
\arguments{
\item{type}{Type of message, either INFO or WARN}

\item{...}{\code{\link{sprintf}} style format string and arguments}
}
\description{
Output a formatted log message. Ellipsis arguments are first passed
to \code{\link{sprintf}} before being printed to console with
\code{\link{print}}.
}
\examples{
log_msg(type = "INFO", "Hello \%s", "world")

info("Information message")

warn("Warning message")

}
